+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - `:primary`: Will always read from a primary node.
     - `:primary_preferred`: Attempt a primary first, then secondary if none available.
     - `:secondary`: Will always read from a secondary node.
+    - `:secondary_preferred`: Attempt a secondary first, then primary if none available.
 
 ## 1.4.3